Добавляет TripController
